Split development-related deps into their own file.
